move the full-run builder out of checks/ and into external_testing/
`checks` is documented as "structural and connectivity checks for lumped-parameter
models" -- LumpedBCVesselCheck and friends, validating a user's model during
generation. A builder that runs a whole study so another repository can test its
reader against the result is not that, and `checks/` is not where anyone would look
for it.

It still has to ship: the wheel contains no `tests/`, and CUFLynx reaches CA through
whatever libcuflynx is installed, so a builder kept in the test suite is one CUFLynx
cannot import. `external_testing` says which of the two it is -- shipped library code
whose only callers are tests outside this repo.

Renaming it now rather than later: v0.5.0 puts this path on PyPI, and CUFLynx pins it
by string through ca_from(). After the tag, moving it is a breaking change across both
repos. CUFLynx's side is updated in the matching PR.
